優化修改

+++ js/app.js
@@ -1,5 +1,4 @@
 const api_path = "pepe1113";
-const token = "AKcTDEUKmxMrKpIUos5xzvZh1aq1";
 const url = `https://hexschoollivejs.herokuapp.com/api/livejs/v1/customer/${api_path}`;
 
 let productList = [];
@@ -9,13 +8,16 @@ const productWrap = document.querySelector(".productWrap");
 const carts = document.querySelector(".js-carts");
 const deleteAllBtn = document.querySelector(".discardAllBtn");
 const orderForm = document.querySelector("#orderInfo");
-const orderInputs = document.querySelectorAll("#orderInfo input");
+const orderInputs = document.querySelectorAll(
+  "input[type=text],input[type=tel],input[type=email]"
+);
 const submitBtn = document.querySelector(".orderInfo-btn");
 
 //初始化
 function init() {
   getProductList();
   getCartList();
+  ValueErrorCheck();
 }
 init();
 
@@ -31,7 +33,6 @@ function getProductList() {
       console.log(err);
     });
 }
-
 //渲染產品列表
 function renderProductList(list) {
   productWrap.innerHTML = "";
@@ -49,7 +50,6 @@ function renderProductList(list) {
 </li>`;
   });
 }
-
 //GET 取得購物車列表
 function getCartList() {
   axios
@@ -83,10 +83,22 @@ function toRenderCartStr(item, index) {
     </div>
     </td>
     <td>NT$${price}</td>
-    <td>${item.quantity}</td>
+    <td>
+    <a href="#">
+    <span class="material-icons" data-num="${item.quantity - 1}" data-id="${
+    item.id
+  }">remove</span></a>
+    <span>${item.quantity}</span>
+    <a href="#">
+    <span class="material-icons" data-num="${item.quantity + 1}" data-id="${
+    item.id
+  }">add</span></a>
+    </td>
     <td>NT$${total}</td>
     <td class="discardBtn">
-        <a href="#" class="material-icons" id="js-clear" data-id=${item.id}> clear </a>
+        <a href="#" class="material-icons" id="js-clear" data-id=${
+          item.id
+        }> clear </a>
     </td>
     </tr>`;
 }
@@ -112,7 +124,6 @@ function addCartList(e) {
     loading();
   }
 }
-
 function postCartList(obj) {
   axios
     .post(`${url}/carts`, obj)
@@ -126,6 +137,34 @@ function postCartList(obj) {
       console.log(err);
     });
 }
+
+//PUT 編輯購物車數量
+function editCartNum(e) {
+  e.preventDefault();
+  let id = e.target.dataset.id;
+  let num = parseInt(e.target.dataset.num);
+  //排除數量為0的報錯
+  if (num === 0) {
+    deleteCartListApi(id);
+    return;
+  }
+
+  axios
+    .patch(`${url}/carts`, {
+      data: {
+        id: id,
+        quantity: num,
+      },
+    })
+    .then(function (res) {
+      getCartList();
+      alert("修改商品數量成功");
+    })
+    .catch(function (err) {
+      console.log(err);
+    });
+}
+carts.addEventListener("click", editCartNum);
 
 //loading page by jquery-loading
 function loading() {
@@ -154,9 +193,12 @@ function deleteAllCartList() {
       alert("購物車已無商品");
     });
 }
-
 deleteAllBtn.addEventListener("click", function (e) {
   e.preventDefault();
+  if (cartList.length == 0) {
+    alert("購物車沒有商品");
+    return;
+  }
   deleteAllCartList();
   loading();
 });
@@ -171,7 +213,6 @@ function deleteCartList(e) {
     deleteCartListApi(productId);
   }
 }
-
 function deleteCartListApi(id) {
   axios
     .delete(`${url}/carts/${id}`)
@@ -179,9 +220,10 @@ function deleteCartListApi(id) {
       cartList = res.data.carts;
       finalTotal = res.data.finalTotal;
       renderCartList();
-      alert("已刪除商品！");
+      alert("刪除商品成功");
     })
     .catch(function (err) {
+      alert("刪除商品失敗");
       console.log(err);
     });
 }
@@ -198,7 +240,11 @@ const constraints = {
       message: "為必填項目",
     },
     numericality: {
-      message: "^請填寫正確電話號碼",
+      message: "^請填寫正確電話格式",
+    },
+    length: {
+      minimum: 8,
+      tooShort: "^應超過8碼數字",
     },
   },
   Email: {
@@ -215,22 +261,22 @@ const constraints = {
     },
   },
 };
-
 function ValueErrorCheck() {
   orderInputs.forEach(function (item) {
-    item.addEventListener("change", function (e) {
-      item.nextElementSibling.textContent = "";
-      let errors = validate(orderForm, constraints);
-      item.nextElementSibling.textContent = errors[this.name];
+    item.addEventListener("blur", function (e) {
+      item.nextElementSibling.textContent = " ";
+      let errors = validate(orderForm, constraints) || "";
+      item.nextElementSibling.textContent =
+        errors[e.target.getAttribute("name")];
     });
   });
 }
-ValueErrorCheck();
 
 function checkErrors() {
-  let errors = validate(orderForm, constraints);
+  let errors = validate(orderForm, constraints) || "";
   let list = Object.keys(errors);
   list.forEach(function (item) {
+    //.item為填單未符合標準出現的提示文字
     document.querySelector(`.${item}`).textContent = errors[item];
   });
 }
@@ -247,6 +293,10 @@ const pay = document.querySelector("#tradeWay");
 submitBtn.addEventListener("click", addOrder);
 function addOrder(e) {
   e.preventDefault();
+  if (cartList.length == 0) {
+    alert("購物車沒有商品，請加入商品");
+    return;
+  }
   let errors = validate(orderForm, constraints);
   if (errors) {
     checkErrors();
@@ -255,7 +305,6 @@ function addOrder(e) {
   getFormValue();
   addOrderPost(orderData);
 }
-
 function getFormValue() {
   let user = {
     name: formName.value,
@@ -267,7 +316,6 @@ function getFormValue() {
   data.user = user;
   orderData.data = data;
 }
-
 function addOrderPost(obj) {
   axios
     .post(`${url}/orders`, obj)
@@ -280,14 +328,12 @@ function addOrderPost(obj) {
     })
     .catch(function (error) {
       console.log(error);
-      alert("購物車沒有商品哦 Σヽ(ﾟД ﾟ; )ﾉ");
     });
 }
 
 //篩選產品列表
 const productSelect = document.querySelector(".productSelect");
 productSelect.addEventListener("change", changeSelectValue);
-
 function changeSelectValue(e) {
   let data = [];
   productList.forEach(function (item) {
